# perf: optimize parsing in Connection API