--- a/google-cloud-spanner/src/main/java/com/google/cloud/spanner/connection/AbstractStatementParser.java
+++ b/google-cloud-spanner/src/main/java/com/google/cloud/spanner/connection/AbstractStatementParser.java
@@ -513,7 +513,7 @@ public abstract class AbstractStatementParser {
     return parsedStatement.copy(statement, defaultQueryOptions);
   }
 
-  private ParsedStatement internalParse(Statement statement, QueryOptions defaultQueryOptions) {
+  ParsedStatement internalParse(Statement statement, QueryOptions defaultQueryOptions) {
     StatementHintParser statementHintParser =
         new StatementHintParser(getDialect(), statement.getSql());
     ReadQueryUpdateTransactionOption[] optionsFromHints = EMPTY_OPTIONS;
@@ -523,16 +523,21 @@ public abstract class AbstractStatementParser {
           statement.toBuilder().replace(statementHintParser.getSqlWithoutClientSideHints()).build();
       optionsFromHints = convertHintsToOptions(statementHintParser.getClientSideStatementHints());
     }
+    // TODO: Qualify statements without removing comments first.
     String sql = removeCommentsAndTrim(statement.getSql());
     ClientSideStatementImpl client = parseClientSideStatement(sql);
     if (client != null) {
       return ParsedStatement.clientSideStatement(client, statement, sql);
-    } else if (isQuery(sql)) {
-      return ParsedStatement.query(statement, sql, defaultQueryOptions, optionsFromHints);
-    } else if (isUpdateStatement(sql)) {
-      return ParsedStatement.update(statement, sql, checkReturningClause(sql), optionsFromHints);
-    } else if (isDdlStatement(sql)) {
-      return ParsedStatement.ddl(statement, sql);
+    } else {
+      String sqlWithoutHints =
+          !sql.isEmpty() && sql.charAt(0) == '@' ? removeStatementHint(sql) : sql;
+      if (isQuery(sqlWithoutHints)) {
+        return ParsedStatement.query(statement, sql, defaultQueryOptions, optionsFromHints);
+      } else if (isUpdateStatement(sqlWithoutHints)) {
+        return ParsedStatement.update(statement, sql, checkReturningClause(sql), optionsFromHints);
+      } else if (isDdlStatement(sqlWithoutHints)) {
+        return ParsedStatement.ddl(statement, sql);
+      }
     }
     return ParsedStatement.unknown(statement, sql);
   }
@@ -935,7 +940,8 @@ public abstract class AbstractStatementParser {
       appendIfNotNull(result, startQuote);
       appendIfNotNull(result, startQuote);
     }
-    while (currentIndex < sql.length()) {
+    int length = sql.length();
+    while (currentIndex < length) {
       char currentChar = sql.charAt(currentIndex);
       if (currentChar == startQuote) {
         if (supportsDollarQuotedStrings() && currentChar == DOLLAR) {
@@ -946,7 +952,7 @@ public abstract class AbstractStatementParser {
             return currentIndex + tag.length() + 2;
           }
         } else if (supportsEscapeQuoteWithQuote()
-            && sql.length() > currentIndex + 1
+            && length > currentIndex + 1
             && sql.charAt(currentIndex + 1) == startQuote) {
           // This is an escaped quote (e.g. 'foo''bar')
           appendIfNotNull(result, currentChar);
@@ -955,7 +961,7 @@ public abstract class AbstractStatementParser {
           continue;
         } else if (isTripleQuoted) {
           // Check if this is the end of the triple-quoted string.
-          if (sql.length() > currentIndex + 2
+          if (length > currentIndex + 2
               && sql.charAt(currentIndex + 1) == startQuote
               && sql.charAt(currentIndex + 2) == startQuote) {
             appendIfNotNull(result, currentChar);
@@ -969,7 +975,7 @@ public abstract class AbstractStatementParser {
         }
       } else if (supportsBackslashEscape()
           && currentChar == BACKSLASH
-          && sql.length() > currentIndex + 1
+          && length > currentIndex + 1
           && sql.charAt(currentIndex + 1) == startQuote) {
         // This is an escaped quote (e.g. 'foo\'bar').
         // Note that in raw strings, the \ officially does not start an escape sequence, but the