This fixes a performance bottleneck I noticed while profiling pgadapter under a TPCC workload.

#### With change

```
Benchmark                                Mode  Cnt       Score      Error  Units
StatementParserBenchmark.isQueryTest    thrpt    5  475330.395 ± 2522.511  ops/s
StatementParserBenchmark.longDmlTest    thrpt    5    3611.682 ±   58.704  ops/s
StatementParserBenchmark.longQueryTest  thrpt    5    5015.354 ±    8.138  ops/s
```

#### Main

```
Benchmark                                Mode  Cnt       Score       Error  Units
StatementParserBenchmark.isQueryTest    thrpt    5  338079.227 ± 26995.410  ops/s
StatementParserBenchmark.longDmlTest    thrpt    5    1153.123 ±     5.604  ops/s
StatementParserBenchmark.longQueryTest  thrpt    5    2011.550 ±    72.433  ops/s
```